v0.0.1后端模型完成

+++ Koole-backend/src/main/java/com/kldo/koolebackend/service/UpdatePostService.java
@@ -1,8 +1,16 @@
 package com.kldo.koolebackend.service;
 
+import com.kldo.koolebackend.converter.UpdatePostConverter;
+import com.kldo.koolebackend.dto.UpdatePostBackDTO;
 import com.kldo.koolebackend.entity.UpdatePost;
+import com.kldo.koolebackend.exception.BusinessException;
+import com.kldo.koolebackend.info.UpdatePostInfo;
 import com.kldo.koolebackend.repository.UpdatePostRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.Caching;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,28 +20,66 @@ public class UpdatePostService {
     @Autowired
     private UpdatePostRepository updatePostRepository;
 
-    public List<UpdatePost> findAll(){
-        return updatePostRepository.findAll();
+    /**
+     * 获取所有更新帖子
+     * @return 更新帖子列表
+     */
+    @Cacheable(value = "updatePostInfoCache",key = "'all'")
+    public List<UpdatePostInfo> findAll(){
+        return  UpdatePostConverter.convertToListInfo(updatePostRepository.findAll());
     }
 
-    public UpdatePost findById(Long id){
-        return updatePostRepository.findById(id).orElse(null);
+    /**
+     * 根据ID获取更新帖子
+     * @param id 更新帖子ID
+     * @return 更新帖子
+     */
+    @Cacheable(value = "updatePostInfoCache",key = "#id")
+    public UpdatePostInfo findById(Long id){
+        UpdatePost updatePost = updatePostRepository.findById(id).orElseThrow(() -> new BusinessException("更新帖子不存在"));
+        return UpdatePostConverter.convertToInfo(updatePost);
     }
 
-    public void create(UpdatePost updatePost){
+    /**
+     * 创建更新帖子
+     * @param updatePostBackDTO 更新帖子返回DTO
+     */
+    @CacheEvict(value = "updatePostInfoCache",key = "'all'")
+    @Transactional
+    public void create(UpdatePostBackDTO updatePostBackDTO) {
+        UpdatePost updatePost = UpdatePostConverter.convertToPost(updatePostBackDTO);
         updatePostRepository.save(updatePost);
     }
 
-    public void update(UpdatePost updatePost){
-        UpdatePost existingUpdatePost = updatePostRepository.findById(updatePost.getPostId()).orElse(null);
-        if(existingUpdatePost != null){
-            existingUpdatePost.setTitle(updatePost.getTitle());
-            existingUpdatePost.setContent(updatePost.getContent());
-            updatePostRepository.save(existingUpdatePost);
-        }
+    /**
+     * 更新更新帖子
+     * @param id 更新帖子ID
+     * @param updatePostBackDTO 更新帖子返回DTO
+     */
+    @Caching(evict = {
+            @CacheEvict(value = "updatePostInfoCache",key = "#id"),
+            @CacheEvict(value = "updatePostInfoCache",key = "'all'")
+    })
+    @Transactional
+    public void update(Long id, UpdatePostBackDTO updatePostBackDTO) {
+        UpdatePost updatePost = updatePostRepository.findById(id).orElseThrow(() -> new BusinessException("更新帖子不存在"));
+        UpdatePostConverter.setPost(updatePostBackDTO, updatePost);
+        updatePostRepository.save(updatePost);
     }
 
+    /**
+     * 删除更新帖子
+     * @param id 更新帖子ID
+     */
+    @Caching(evict = {
+            @CacheEvict(value = "updatePostInfoCache",key = "#id"),
+            @CacheEvict(value = "updatePostInfoCache",key = "'all'")
+    })
+    @Transactional
     public void delete(Long id) {
+        if(!updatePostRepository.existsById(id)){
+            throw new BusinessException("更新帖子不存在");
+        }
         updatePostRepository.deleteById(id);
     }
 }
